backend_api/requirements.txt
Flask==3.0.0
Flask-Cors
Flask-PyMongo
dnspython
python-dotenv
ultralytics
google-genai